+++ src/lax/mod.rs
@@ -74,7 +74,7 @@
 pub mod category;
 pub mod functor;
 pub mod hypergraph;
-pub mod subgraph;
+pub mod matching;
 pub mod mut_category;
 pub mod open_hypergraph;
