+++ tests/lax/mod.rs
@@ -1,3 +1,3 @@
 pub mod eval;
 pub mod hypergraph;
-pub mod subgraph;
+pub mod matching;
